requirements.txt
# Tested with Python 3.13.12
torch==2.11.0
PyYAML==6.0.3
pytest==9.0.2
numpy==2.4.4
matplotlib==3.10.8